Tr Type

+++ db/training.h
@@ -72,8 +72,8 @@ private:
     void positionUpdated(const QGeoPositionInfo& info);
     QGeoPositionInfoSource* m_geoSource;
     QGeoCoordinate m_lastPoint;
-    int m_distanceWithStimulation = 0;
-    int m_distanceWithoutStimulation = 0;
+    double m_distanceWithStimulation = 0;
+    double m_distanceWithoutStimulation = 0;
     int m_speedWithStimulation = 0;
     int m_speedWithoutStimulation = 0;
